cifar_experiment/scripts/requirements.txt
numpy --no-index
tensorflow_gpu --no-index
tensorflow_probability --no-index
